impliment image upload to firebase

+++ app/src/main/java/com/ayoub/electricitybill/model/Bill.kt
@@ -6,5 +6,5 @@ data class Bill(
     val image: String,
     val extra: Double,
     val consumptions: List<Consumption>,
-    val createdAt: Double,
+    val createdAt: Long,
 )
